Implemented brute force solutions for Kangaroo contest
In this update, brute force solutions to various problems from the Kangaroo Math Competition have been introduced. The Kangaroo competition is a reputable international mathematical contest that challenges students with a range of engaging problems. While brute force might not be the most efficient or elegant approach to solving these problems, it's a method that guarantees finding a solution, albeit in a longer time frame. This directory now houses multiple Java files, each tackling a different problem from the competition. This endeavor is not only a means to find solutions but also an exercise in understanding the nature and structure of problems presented in this competition. This commit invites further exploration and optimization of solutions.